src/api_adapters/tg_adapter/requirements.txt
#git+https://github.com/confluentinc/confluent-kafka-python
kafka-python
requests==2.32.3
